add user permissions for games

+++ levelupapi/views/games.py
@@ -6,26 +6,33 @@ from rest_framework.decorators import renderer_classes
 from rest_framework.serializers import Serializer
 from rest_framework.viewsets import ViewSet
 from rest_framework.response import Response
+from rest_framework import permissions
 from rest_framework import serializers
 from levelupapi.models import Game, GameType, Gamer
 
-class GameSerializer(serializers.HyperlinkedModelSerializer):
-    """JSON serializer for games
-
-    Arguments:
-        serializer type
+class IsOwnerOrReadOnly(permissions.BasePermission):
     """
-    class Meta:
-        model = Game
-        url = serializers.HyperlinkedIdentityField(
-            view_name='game',
-            lookup_field='id'
-        )
-        fields = ('id', 'url', 'title', 'maker', 'number_of_players', 'skill_level', 'game_type')
-        depth = 1
+    Object-level permission to only allow owners of an object to edit it.
+    Assumes the model instance has an `owner` attribute.
+    """
+
+    def has_object_permission(self, request, view, obj):
+        # Read permissions are allowed to any request,
+        # so we'll always allow GET, HEAD or OPTIONS requests.
+        if request.method in permissions.SAFE_METHODS:
+            return True
+
+        # Instance must have an attribute named `owner`.
+        return obj.gamer.user == request.user
 
 class GamesViewSet(ViewSet):
     """Level up games"""
+
+    # Tells django which permission type to use
+    permission_classes = [ IsOwnerOrReadOnly ]
+
+    # Requires a queryset property so the view knows about the model and permissions
+    queryset = Game.objects.none()
 
     def create(self, request):
         """Handle POST operations
@@ -84,6 +91,7 @@ class GamesViewSet(ViewSet):
         gamer = Gamer.objects.get(user=request.auth.user)
 
         game = Game.objects.get(pk=pk)
+        self.check_object_permissions(request, game)
         game.title = request.data["title"]
         game.maker = request.data["maker"]
         game.number_of_players = request.data["numberOfPlayers"]
@@ -105,6 +113,7 @@ class GamesViewSet(ViewSet):
         
         try:
             game = Game.objects.get(pk=pk)
+            self.check_object_permissions(request, game)
             game.delete()
 
             return Response({}, status=status.HTTP_204_NO_CONTENT)
@@ -132,3 +141,19 @@ class GamesViewSet(ViewSet):
         serializer = GameSerializer(
             games, many=True, context={'request': request})
         return Response(serializer.data)
+
+
+class GameSerializer(serializers.HyperlinkedModelSerializer):
+    """JSON serializer for games
+
+    Arguments:
+        serializer type
+    """
+    class Meta:
+        model = Game
+        url = serializers.HyperlinkedIdentityField(
+            view_name='game',
+            lookup_field='id'
+        )
+        fields = ('id', 'url', 'title', 'maker', 'number_of_players', 'skill_level', 'game_type')
+        depth = 1
